transferring project to docker. Absolute mindfuck. Next to do: getting the supabase realtime to work properly

+++ stored_functions/authenticate.sql
@@ -24,7 +24,7 @@ begin
 		'role', account_data.role,
 		'profile_id', account_data.id,
 		'token_version', account_data.token_version,
-		'exp', extract(epoch from now() + interval '7 days')), 'thisstringissoverysecretextrachars');
+		'exp', extract(epoch from now() + interval '7 days')::bigint), 'thisstringissoverysecretextrachars');
 
 	return jwt;
 end;
